backend/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
dj-database-url==3.0.1
psycopg2-binary==2.9.10
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
mysql-connector-python==9.5.0
pillow==12.1.0
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.2
